requirements.txt
requests
python-dotenv
urllib3
snowflake-connector-python
